JP-682: Revision of persistence of pluggable classes

+++ comu/com.energyict.comserver.comserver.core.rest/src/main/java/com/energyict/mdc/rest/impl/comserver/ComPortPoolInfo.java
@@ -26,7 +26,7 @@ public abstract class ComPortPoolInfo<T extends ComPortPoolShadow> {
     @XmlJavaTypeAdapter(ComPortTypeAdapter.class)
     public ComPortType type;
     public List<InboundComPortInfo> inboundComPorts;
-    public int discoveryProtocolPluggableClassId;
+    public long discoveryProtocolPluggableClassId;
     public List<OutboundComPortInfo> outboundComPorts;
     public TimeDurationInfo taskExecutionTimeout;
